refactor: resolve circular dependency in main process
Extracted window management logic into a new `src/main/windowManager.ts` file to resolve a circular dependency between `main.ts`, `watcher.ts`, and `storage.ts`.

- `main.ts` now imports `setMainWindow` and `getMainWindow` from `windowManager.ts`.
- `watcher.ts` and `storage.ts` now import `sendNotification` and `sendUnmatchedFiles` from `windowManager.ts`.
